+++ assemblyadapter-list/src/main/java/com/github/panpf/assemblyadapter/list/expandable/AssemblySingleDataExpandableListAdapter.kt
@@ -18,9 +18,7 @@ package com.github.panpf.assemblyadapter.list.expandable
 import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseExpandableListAdapter
-import com.github.panpf.assemblyadapter.AssemblyAdapter
-import com.github.panpf.assemblyadapter.Item
-import com.github.panpf.assemblyadapter.ItemFactory
+import com.github.panpf.assemblyadapter.*
 import com.github.panpf.assemblyadapter.internal.ItemFactoryStorage
 import com.github.panpf.assemblyadapter.list.R
 
@@ -229,14 +227,22 @@ open class AssemblySingleDataExpandableListAdapter<GROUP_DATA : Any, CHILD_DATA>
 
 
     override fun getItemFactoryByPosition(position: Int): ItemFactory<*> {
-        val groupCount = groupCount
-        if (position < 0 || position >= groupCount) {
-            throw IndexOutOfBoundsException("Index: $position, Size: $groupCount")
-        }
+        val data = getGroup(position)
         return itemFactoryStorage.getItemFactoryByData(
-            data!!, "ItemFactory", "AssemblySingleDataExpandableListAdapter", "itemFactoryList"
+            data, "ItemFactory", "AssemblySingleDataExpandableListAdapter", "itemFactoryList"
         )
     }
 
-    // todo add fun getItemFactoryByPosition(groupPosition: Int, childPosition: Int): ItemFactory<*>
+    /**
+     * Get the ItemFactory of the specified [childPosition]
+     *
+     * @throws IndexOutOfBoundsException If the [groupPosition] or [childPosition] is out of range
+     * @throws NotFoundMatchedItemFactoryException No ItemFactory can match the data corresponding to [childPosition]
+     */
+    fun getItemFactoryByChildPosition(groupPosition: Int, childPosition: Int): ItemFactory<*> {
+        val data = getChild(groupPosition, childPosition) ?: Placeholder
+        return itemFactoryStorage.getItemFactoryByData(
+            data, "ItemFactory", "AssemblyExpandableListAdapter", "itemFactoryList"
+        )
+    }
 }